refactor(api): stages → src/api modules (Effort 1)
Mechanical move of the stages feature from src/hooks/queries/stages to
src/api/stages, adding queryOptions factories for the query hooks. No behavior change.
